refactor: OpenChecker 인터페이스 생성 및 테스트 리팩토링
